Changed structure of files for settings
