+++ custom-cookie-message.php
@@ -20,9 +20,12 @@ define( 'CUSTOM_COOKIE_MESSAGE_PLUGIN_URL', plugins_url( '', __FILE__ ) . '/' );
 define( 'CUSTOM_COOKIE_MESSAGE_PLUGIN_SLUG', 'custom-cookie-message-list' );
 
 require_once( 'includes/ac-cookie-message.php' );
+require_once( 'includes/settings.php' );
 
 add_action( 'plugins_loaded', function () {
 	$GLOBALS['CustomCookie'] = new AC_Custom_Cookie_Message();
+	$GLOBALS['Settings'] = new AC_Custom_Cookie_Message_Admin_Settings();
+
 } );
 
 register_activation_hook( __FILE__, array( 'AC_Custom_Cookie_Message', 'plugin_activation' ) );
